\name{permDNA}
\alias{permDNA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Permutation of DNA sequences
}
\description{
Returns all possible permutations of DNA sequences of a given length.
}

\usage{
permDNA(K)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{K}{Length of required sequences}
}
\details{

}
\value{
A character vector giving all possible permutations of DNA sequences of length \code{K}.
}
\references{

}
\author{
Samuel Brown <s_d_j_brown@hotmail.com>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
permDNA(1)
permDNA(2)
permDNA(3)
permDNA(4)
permDNA(5)
}
