\name{pindel}
\alias{pindel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{DNA distance model incorporating alignment gaps}

\description{Calculates the p-distance between sequences, incorporating gaps as a fifth character}

\usage{
pindel(DNAbin, proportion = TRUE, indelmod = "indel")
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{DNAbin}{A `DNAbin' object.}
  \item{proportion}{Something I'll need to figure out later}
  \item{indelmod}{Something to do with indels.}
}

\details{
}

\value{
A distance matrix
}

\references{
}


\author{
Samuel Brown <s_d_j_brown@hotmail.com>
}

\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link[ape]{dist.dna}}
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

\examples{

}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

